requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
pydantic[email]
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==43.0.1
hyperlink==21.0.0
idna
incremental
isodate
logzero
pycparser
pyotp
pyparsing
python-dateutil
pytz
rdflib
rdflib-jsonld
requests
simplejson
six
smartapi-python
Twisted
txaio
typing_extensions
urllib3
websocket-client
zope.interface
logzero
smartapi-python
pandas
